[SGP-16069] fix unittests
This fix works by defaulting to a bit of the old locust usage.

Essentially, if no userlcass was specified in the command line for
the locust deployment or not userclass was specified in the command line
when running in headless, all users would be run. Default to that so that
the unittests can work either way.